initial version of CU_SpMM::GIN layer

+++ cuda/cuBlasUtil.h
@@ -1,3 +1,3 @@
 void gpu_blas_mmul(const float *A, const float *B, float *C, int m, int k, int n, bool transA, bool transB, float alpha, float beta);
 
-__global__ void initIdentityGPU(int **devMatrix, int numR, int numC);
+__global__ void initIdentityGPU(float **devMatrix, int numR, int numC);
